requirements.txt
# --------- pytorch --------- #
# Install a CUDA build that matches the host NVIDIA driver (see `nvidia-smi` "CUDA Version").
# Example for CUDA 12.x drivers (e.g. Lambda when smi shows 12.8): install from PyTorch wheels
# before `pip install -r requirements.txt`, or in one line:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
# Plain `pip install torch` may resolve to a newer CUDA wheel (e.g. +cu130) than the driver supports.
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550
tb-nightly

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
path
matplotlib

natsort